refactor: rename `plugin` option for better semantic
